+++ FinalProject/FinalProject.vcxproj
@@ -141,14 +141,14 @@
   <ItemGroup>
     <ClCompile Include="Activity.cpp" />
     <ClCompile Include="Main.cpp" />
-    <ClCompile Include="AbstractMenu.cpp" />
+    <ClCompile Include="Menu.cpp" />
     <ClCompile Include="Rank.cpp" />
     <ClCompile Include="TJ.cpp" />
     <ClCompile Include="User.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Activity.h" />
-    <ClInclude Include="AbstractMenu.h" />
+    <ClInclude Include="Menu.h" />
     <ClInclude Include="Rank.h" />
     <ClInclude Include="TJ.h" />
     <ClInclude Include="User.h" />
